requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2
django-environ==0.4.5
django-froala-editor==3.2.2
djangorestframework==3.12.4
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
wincertstore==0.2
